move history session to home
