+++ components/common/List.js
@@ -20,19 +20,18 @@ const List = memo(({ name }) => {
                 <tr key={`item-` + index} className="hover:bg-gray-100 border border-transparent rounded-md text-lg">
                   <td className="pl-4 py-2 rounded-tl-md rounded-bl-md">{item.name}</td>
                   <td className="pr-4 py-2 w-10 rounded-tr-md rounded-br-md">
-                   {!(sessionStorage.getItem('sid')) ? (
-
-                    <button
-                      type="button"
-                      onClick={() => {
-                        const newValue = watch(name).slice()
-                        newValue.splice(index, 1)
-                        toast.success('Xóa thành công', { position: 'top-right' })
-                        onChange(newValue)
-                      }}
-                    >
-                      <TrashIcon className="w-6" />
-                    </button>
+                    {!sessionStorage.getItem('sid') ? (
+                      <button
+                        type="button"
+                        onClick={() => {
+                          const newValue = watch(name).slice()
+                          newValue.splice(index, 1)
+                          toast.success('Xóa thành công', { position: 'top-right' })
+                          onChange(newValue)
+                        }}
+                      >
+                        <TrashIcon className="w-6" />
+                      </button>
                     ) : null}
                   </td>
                 </tr>
