+++ consts/urls.js
@@ -1,4 +1,5 @@
 export default {
+  LOGIN: '/login',
   HOME: '/',
   SESSIONS_CREATE: '/sessions/create',
   SESSIONS: '/sessions/detail',
