+++ style/home.css
@@ -5,6 +5,7 @@
   background-position: center;
   background-repeat: no-repeat;
   min-height: 100%;
+  padding-top: 20px;
 }
 .bg-image1 {
   /* The image used */
@@ -16,16 +17,18 @@
   min-width: 100%;
   min-height: 100%;
   overflow-y: auto;
+  padding-top: 20px;
 }
 .bg-image2 {
   /* The image used */
-  background: url('/assets/images/bg25.png') no-repeat; 
+  background: url('/assets/images/bg25.png') no-repeat;
   background-attachment: local;
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
   min-height: 100%;
   min-width: 100%;
+  padding-top: 20px;
 }
 body {
   overflow-x: hidden;
